extract tmp file handling
